Change the way query string is handled.
Each component used to have a declarative way to handle query strings

class Component(...):
	_url_params: dict[str, str]

but that's not flexible enough or parctical in real life, after trying
to use it. I'm chaging that to a mutable dict in `self.reactor.params`
where you can modify or access query paramters to restore or persist
pieces of your component state.

Signed-off-by: Eddy del Valle Pino <eddy@edelvalle.me>