Go back to idiomorph and do not use custom elements
While using idiomorph custom elements where a pain, because the elements do not get added or removed but really morphed.
Using morphdom triggers inncesary add/remove of HTML elements causing a component to send Leave and Join in the same render.

This solution uses idiomorph and when the DOM is modified it is scanned for new elements and old elements that should be removed.